Increase timeout when downloading sources
The default timeout (10 sec) sometimes activates for larger
sources such as VTK-m. Avoid this problem by increasing the
timeout to 100 sec.

I'm not 100% sure this is working correctly. I think this
is causing the environment to overwrite the default value.
It wasn't working before and it worked the next time I tried
it, but intermittent successes happen.